Fix errors in deepspeed>=0.16.5 (#3963)
There was a quite major change made in deepspeed 0.16.5 where they started to apply forward/backward pre- and post-processing not only to the deepspeed engine object but also to the original model itself.
This causes subtle disruption with aimet-torch and makes deepspeed runtime misinterpret quantizer parameters as "external" parameters.
This commit changes the parameter access pattern of aimet-torch quantized modules in a deepspeed-friendly fashion so that it doesn't trigger such misinterpretation anymore

Signed-off-by: Kyunggeun Lee <quic_kyunggeu@quicinc.com>